env: dev: bashrc: Add helper functions for EC2.
To make it easy to launch AL2023/AL2/ALAMI, ral23/ral2/ral1 can be
used with name for the instance:

  ral23 al23-test

The aarch64 version is ral23a and ral2a.

The default instance type is c7i.4xlarge and c7g.4xlarge, but it can
be overridden by setting INSTANCE_TYPE:

  INSTANCE_TYPE ral23 al23-test

Also, AMI ID can be directly used with rec2 and rec2a:

  rec2 $INSTANCE_NAME $AMI_ID

Another helper function is ec2-serial, which connects to an instance
with the specified instance ID using serial console.

Note that these variables must be set before use.

  * KEYPAIR_NAME=""
  * SUBNET_ID=""
  * SECURITY_GROUP_ID=""

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>